refactor: move TransactionalDbmsSupportAwareDataSourceProxy to dbms module
