+++ core/src/main/java/nl/nn/adapterframework/jndi/JndiDataSourceFactory.java
@@ -25,12 +25,13 @@ import javax.sql.DataSource;
 
 import org.springframework.jdbc.datasource.TransactionAwareDataSourceProxy;
 
+import nl.nn.adapterframework.dbms.TransactionalDbmsSupportAwareDataSourceProxy;
 import nl.nn.adapterframework.jdbc.IDataSourceFactory;
 import nl.nn.adapterframework.util.AppConstants;
 
 /**
  * Factory through which (TX-enabled) DataSources can be retrieved.
- * 
+ *
  * Already created DataSources are stored in a ConcurrentHashMap.
  * Every DataSource can be augmented before it is added.
  */
